refactor: rename interactive-lab-callout
