+++ src/components/interactive-lab-callout/docs.mdx
@@ -1,5 +1,5 @@
 ---
-componentName: LabCallout
+componentName: InteractiveLabCallout
 ---
 
 # Interactive Lab Callout
@@ -8,16 +8,9 @@ This component showcases a callout to draw attention to the interactive lab feat
 
 ## Playground
 
-<!--
-Note: component is actually called InteractiveLabCallout.
-But the name mismatch makes Swingset angry.
-I think we'd need to rename the component folder in order
-to make Swingset happy.
--->
-
 <LiveComponent>{`
 <InstruqtProvider labId="hashicorp-learn/tracks/read-secrets-from-vault-using-vault-agent?token=em_CzIllfgt8ushWfcm">
-    <LabCallout />
+    <InteractiveLabCallout />
 </InstruqtProvider>
 `}</LiveComponent>
