+++ src/components/interactive-lab-callout/img/interactive-callout-visual.png
